+++ src/dynamodb.ts
@@ -5,7 +5,7 @@ import { RemovableResource } from "./resource";
 
 /**
  * A test construct for a DynamoDB Table
- * @see {@link https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_dynamodb.Table.html}
+ * @see {@link https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_dynamodb.CfnTable.html}
  */
 export class DynamoDBTable extends RemovableResource {
   constructor(template: AdvancedTemplate, props?: any) {
@@ -18,7 +18,7 @@ export class DynamoDBTable extends RemovableResource {
    * @returns 
    */
   public withTableName(name: string) {
-    this.setProperty('TableName', Match.stringLikeRegexp(name));
+    this.withProperty('TableName', Match.stringLikeRegexp(name));
     return this;
   }
 
@@ -29,12 +29,12 @@ export class DynamoDBTable extends RemovableResource {
    * @returns 
    */
   public withKey(key: string, type?: AttributeType) {
-    this.setProperty('KeySchema', Match.arrayWith([
+    this.withProperty('KeySchema', Match.arrayWith([
       Match.objectLike({
         AttributeName: Match.stringLikeRegexp(key),
       }),
     ]));
-    this.setProperty('AttributeDefinitions', Match.arrayWith([
+    this.withProperty('AttributeDefinitions', Match.arrayWith([
       Match.objectEquals({
         AttributeName: Match.stringLikeRegexp(key),
         AttributeType: type || AttributeType.STRING,