+++ src/route53.ts
@@ -1,5 +1,5 @@
-import { Match } from "aws-cdk-lib/assertions";
-import { CfnRecordSet, RecordType } from "aws-cdk-lib/aws-route53";
+import { Match, Matcher } from "aws-cdk-lib/assertions";
+import { CfnHostedZone, CfnRecordSet, RecordType } from "aws-cdk-lib/aws-route53";
 import { AdvancedMatcher } from "./advanced-matcher";
 import { AdvancedTemplate } from "./advanced-template"
 import { CloudFrontDistribution } from "./cloudfront";
@@ -7,7 +7,28 @@ import { RemovableResource } from "./resource";
 
 /**
  * A test construct representing a Route53 RecordSet.
- * @see {@link https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_route53.RecordSet.html}
+ * @see {@link https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_route53.CfnHostedZone.html}
+ */
+export class Route53HostedZone extends RemovableResource {
+  constructor(template: AdvancedTemplate, props?: any) {
+    super(CfnHostedZone.CFN_RESOURCE_TYPE_NAME, template, props);
+  }
+
+  /**
+   * Sets a matching name for the HostedZone
+   * @param name Either the whole or a partial name of the record
+   * @returns 
+   */
+  public withName(name: string) {
+    this.withProperty('Name', Match.stringLikeRegexp(name));
+    return this;
+  }
+}
+
+
+/**
+ * A test construct representing a Route53 RecordSet.
+ * @see {@link https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_route53.CfnRecordSet.html}
  */
 export class Route53RecordSet extends RemovableResource {
   constructor(template: AdvancedTemplate, props?: any) {
@@ -21,7 +42,7 @@ export class Route53RecordSet extends RemovableResource {
    * @returns 
    */
   public withRecordType(recordType: RecordType) {
-    this.setProperty('Type', recordType);
+    this.withProperty('Type', recordType);
     return this;
   }
 
@@ -31,17 +52,45 @@ export class Route53RecordSet extends RemovableResource {
    * @returns 
    */
   public withName(name: string) {
-    this.setProperty('Name', Match.stringLikeRegexp(name));
+    this.withProperty('Name', Match.stringLikeRegexp(name));
     return this;
   }
 
   /**
    * Sets a matching hosted zone ID for the RecordSet
-   * @param zoneId Either the whole or a partial zone ID the record belongs to
+   * @param zone Either the hosted zone resource, a matcher or a matching zone Id.
    * @returns 
    */
-  public inHostedZone(zoneId: string) {
-    this.setProperty('HostedZoneId', zoneId);
+  /*
+    TODO: add mapping
+
+    "HostedZoneId": {
+      "Fn::FindInMap": [
+        "AWSCloudFrontPartitionHostedZoneIdMap",
+        {
+          "Ref": "AWS::Partition"
+        },
+        "zoneId"
+      ]
+    }
+    
+    "Mappings": {
+      "AWSCloudFrontPartitionHostedZoneIdMap": {
+        "aws": {
+          "zoneId": "Z2FDTNDATAQYW2"
+        },
+        "aws-cn": {
+          "zoneId": "Z3RFFRIM2A3IF5"
+        }
+      }
+    },
+  */
+  public inHostedZone(zone: Route53HostedZone | Matcher) {
+    if (zone instanceof Route53HostedZone) {
+      this.withProperty('HostedZoneId', zone.ref);
+    } else {
+      this.withProperty('HostedZoneId', zone);
+    }
     return this;
   }
 
@@ -51,7 +100,7 @@ export class Route53RecordSet extends RemovableResource {
    * @returns 
    */
   public withAliasToCloudFront(distribution: CloudFrontDistribution) {
-    this.setProperty('AliasTarget', Match.objectLike({
+    this.withProperty('AliasTarget', Match.objectLike({
       DNSName: AdvancedMatcher.fnGetAtt(distribution.id, "DomainName"),
     }));
     return this;
@@ -62,9 +111,19 @@ export class Route53RecordSet extends RemovableResource {
    * @returns 
    */
   public withAliasToS3() {
-    this.setProperty('AliasTarget', Match.objectLike({
+    this.withProperty('AliasTarget', Match.objectLike({
       DNSName: `s3-website.${this.template.region}.amazonaws.com`,
     }));
+    return this;
+  }
+
+  /**
+   * Sets matching resource records for the RecordSet
+   * @param records The list of records
+   * @returns 
+   */
+  public withResourceRecords(records: string[]) {
+    this.withProperty('ResourceRecords', Match.arrayWith(records));
     return this;
   }
 }