+++ tests/cloudfront.test.ts
@@ -14,7 +14,7 @@ describe("CloudFront", () => {
     });
     template = AdvancedTemplate.fromStack(stack);
 
-    template.debug();
+    // template.debug();
 
     distribution = template.cloudFrontDistribution();
   });