+++ tests/stacks/custom.stack.ts
@@ -1,4 +1,5 @@
-import { Stack, StackProps } from "aws-cdk-lib";
+import { CfnCondition, CfnMapping, CfnOutput, Fn, Stack, StackProps } from "aws-cdk-lib";
+import { CfnBucket } from "aws-cdk-lib/aws-s3";
 import { AwsCustomResource, AwsCustomResourcePolicy, PhysicalResourceId } from "aws-cdk-lib/custom-resources";
 import { Construct } from "constructs";
 
@@ -8,7 +9,7 @@ export class TestCustomStack extends Stack {
 
     // CUSTOM RESOURCE
 
-    new AwsCustomResource(this, id + 'Custom', {
+    const custom = new AwsCustomResource(this, id + 'Custom', {
       policy: AwsCustomResourcePolicy.fromSdkCalls({
         resources: AwsCustomResourcePolicy.ANY_RESOURCE,
       }),
@@ -30,6 +31,31 @@ export class TestCustomStack extends Stack {
         region: props.env.region,
         physicalResourceId: PhysicalResourceId.of('CustomParameter'),
       },
+    });
+
+    // CONDITION
+
+    const bucket = new CfnBucket(this, id + 'Bucket', {
+      bucketName: 'bucket',
+    });
+    bucket.cfnOptions.condition = new CfnCondition(this, id + 'Condition', {
+      expression: Fn.conditionEquals(true, true),
+    });
+
+    // MAPPING
+
+    new CfnMapping(this, id + 'Mapping', {
+      mapping: {
+        'eu-central-1': {
+          regionName: 'Europe (Frankfurt)'
+        }
+      }
+    });
+
+    // OUTPUT
+
+    new CfnOutput(this, 'CustomARN', {
+      value: custom.node.id,
     });
   }
 }